multi-channel-agent/requirements.txt
anima-labs>=0.1.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
